backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic==1.16.2",
    "asyncpg>=0.31.0",
    "cloudinary>=1.41.0",
    "fastapi==0.115.13",
    "httpx>=0.28.1",
    "numpy>=2.0.0",
    "psycopg2-binary==2.9.10",
    "pydantic==2.11.7",
    "python-dotenv>=1.2.1",
    "python-json-logger==3.3.0",
    "sentence-transformers>=3.0.0",
    "sqlalchemy==2.0.41",
    "uvicorn==0.34.3",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "ipykernel>=7.1.0",
    "pylint>=4.0.4",
    "pytest>=9.0.2",
]
